pyproject.toml
[project]
name = "study-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markdown>=3.8",
    "openai>=1.78.1",
    "playwright>=1.52.0",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
]
